代码修改

+++ src/main/java/com/fooww/research/deviation/ItemDeviation.scala
@@ -22,7 +22,7 @@ object ItemDeviation {
     })
     val itemItem_deviation = itemItem_score.reduceByKey(_+_)
     val resultRDD = itemItem_deviation.map(f=>{
-      (f._1._1,f._1._2,f._2)
+      (f._1._1,f._1._2,f._2.toDouble)
     })
     resultRDD
   }
